create GetCats and GetDogs components
