+++ src/components/CatActivities.js
@@ -10,7 +10,7 @@ import {
 } from "@mui/material";
 import ArrowForwardIosSharpIcon from "@mui/icons-material/ArrowForwardIosSharp";
 
-export default function CatActivities(props) {
+export default function CatActivities() {
   return (
     <Container>
       <Typography
